requirements.txt
streamlit
firecrawl-py
pydantic
python-dotenv
psycopg2-binary
sqlalchemy==2.0.35
pandas
plotly
BeautifulSoup4